TODO tests/xtimer_mutex_lock_timeout: minimal xtimer_mutex_lock_timeout test
This test checks if the function works when the timeout is smaller than XTIMER_BACKOFF and the mutex is already locked.
This means the timer will spin and the timer will shoot before the mutex lock was called.
Then the mutex lock gets called and the timer will not remove the thread from the mutex.
Checking if this case will get handled correctly.