xtimer/xtimer.c: _mutex_remove_thread_from_waiting_queue pointer to return value
The function will now get a int pointer as parameter and write into the int (the pointer is pointing to)
whether or not the thread was unlocked.

This is because if this function is called by a thread and the unlocked thread has a higher priority
the return value cannot be used. This way the unlocked thread can now use the return value
(if he has the pointer to the value)